test selection Submenu
This code demonstrate that when we need to test the menu and submenu functionality works appropriately or not. When we hover on the menu, the submenu should get visible. Likewise when we hover on the SubMenu then the Sub-SubMenu gets displayed and we can use .click() method to navigate to that page.